requirements.txt
sqlalchemy
requests
requests
numpy
pandas
flask 
data
psycopg2-binary
pandas
sklearn
python-decouple